Test replaced by program.cs
